+++ src/main/java/com/clubber/ClubberServer/global/infrastructure/outer/exception/config/DiscordFeignConfig.java
@@ -1,4 +1,4 @@
-package com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.config;
+package com.clubber.ClubberServer.global.infrastructure.outer.exception.config;
 
 import feign.Client;
 import org.springframework.context.annotation.Bean;
